rename serde to ipc
